GODRIVER-3215 Fix default auth source for auth specified via ClientOptions.
