+++ mongo/client.go
@@ -26,7 +26,6 @@ import (
 	"go.mongodb.org/mongo-driver/mongo/writeconcern"
 	"go.mongodb.org/mongo-driver/x/bsonx/bsoncore"
 	"go.mongodb.org/mongo-driver/x/mongo/driver"
-	"go.mongodb.org/mongo-driver/x/mongo/driver/auth"
 	"go.mongodb.org/mongo-driver/x/mongo/driver/mongocrypt"
 	mcopts "go.mongodb.org/mongo-driver/x/mongo/driver/mongocrypt/options"
 	"go.mongodb.org/mongo-driver/x/mongo/driver/operation"
@@ -211,43 +210,16 @@ func NewClient(opts ...*options.ClientOptions) (*Client, error) {
 		clientOpt.SetMaxPoolSize(defaultMaxPoolSize)
 	}
 
-	if clientOpt.Auth != nil {
-		var oidcMachineCallback auth.OIDCCallback
-		if clientOpt.Auth.OIDCMachineCallback != nil {
-			oidcMachineCallback = func(ctx context.Context, args *driver.OIDCArgs) (*driver.OIDCCredential, error) {
-				cred, err := clientOpt.Auth.OIDCMachineCallback(ctx, convertOIDCArgs(args))
-				return (*driver.OIDCCredential)(cred), err
-			}
-		}
-
-		var oidcHumanCallback auth.OIDCCallback
-		if clientOpt.Auth.OIDCHumanCallback != nil {
-			oidcHumanCallback = func(ctx context.Context, args *driver.OIDCArgs) (*driver.OIDCCredential, error) {
-				cred, err := clientOpt.Auth.OIDCHumanCallback(ctx, convertOIDCArgs(args))
-				return (*driver.OIDCCredential)(cred), err
-			}
-		}
-
-		// Create an authenticator for the client
-		client.authenticator, err = auth.CreateAuthenticator(clientOpt.Auth.AuthMechanism, &auth.Cred{
-			Source:              clientOpt.Auth.AuthSource,
-			Username:            clientOpt.Auth.Username,
-			Password:            clientOpt.Auth.Password,
-			PasswordSet:         clientOpt.Auth.PasswordSet,
-			Props:               clientOpt.Auth.AuthMechanismProperties,
-			OIDCMachineCallback: oidcMachineCallback,
-			OIDCHumanCallback:   oidcHumanCallback,
-		}, clientOpt.HTTPClient)
-		if err != nil {
-			return nil, err
-		}
+	client.authenticator, err = topology.NewAuthenticator(clientOpt.Auth, clientOpt.HTTPClient)
+	if err != nil {
+		return nil, fmt.Errorf("error creating authenticator: %w", err)
 	}
 
 	cfg, err := topology.NewConfigWithAuthenticator(clientOpt, client.clock, client.authenticator)
-
 	if err != nil {
 		return nil, err
 	}
+
 	client.serverAPI = topology.ServerAPIFromServerOptions(cfg.ServerOpts)
 
 	if client.deployment == nil {
@@ -264,19 +236,6 @@ func NewClient(opts ...*options.ClientOptions) (*Client, error) {
 	}
 
 	return client, nil
-}
-
-// convertOIDCArgs converts the internal *driver.OIDCArgs into the equivalent
-// public type *options.OIDCArgs.
-func convertOIDCArgs(args *driver.OIDCArgs) *options.OIDCArgs {
-	if args == nil {
-		return nil
-	}
-	return &options.OIDCArgs{
-		Version:      args.Version,
-		IDPInfo:      (*options.IDPInfo)(args.IDPInfo),
-		RefreshToken: args.RefreshToken,
-	}
 }
 
 // Connect initializes the Client by starting background monitoring goroutines.
