+++ x/mongo/driver/topology/topology_options.go
@@ -7,10 +7,10 @@
 package topology
 
 import (
+	"context"
 	"crypto/tls"
 	"fmt"
 	"net/http"
-	"strings"
 	"time"
 
 	"go.mongodb.org/mongo-driver/event"
@@ -71,31 +71,76 @@ func newLogger(opts *options.LoggerOptions) (*logger.Logger, error) {
 	return log, nil
 }
 
-// NewConfig will translate data from client options into a topology config for building non-default deployments.
-func NewConfig(co *options.ClientOptions, clock *session.ClusterClock) (*Config, error) {
-	// Auth & Database & Password & Username
-	if co.Auth != nil {
-		cred := &auth.Cred{
-			Username:    co.Auth.Username,
-			Password:    co.Auth.Password,
-			PasswordSet: co.Auth.PasswordSet,
-			Props:       co.Auth.AuthMechanismProperties,
-			Source:      co.Auth.AuthSource,
-		}
-		mechanism := co.Auth.AuthMechanism
-		authenticator, err := auth.CreateAuthenticator(mechanism, cred, co.HTTPClient)
-		if err != nil {
-			return nil, err
-		}
-		return NewConfigWithAuthenticator(co, clock, authenticator)
+// convertOIDCArgs converts the internal *driver.OIDCArgs into the equivalent
+// public type *options.OIDCArgs.
+func convertOIDCArgs(args *driver.OIDCArgs) *options.OIDCArgs {
+	if args == nil {
+		return nil
 	}
-	return NewConfigWithAuthenticator(co, clock, nil)
+	return &options.OIDCArgs{
+		Version:      args.Version,
+		IDPInfo:      (*options.IDPInfo)(args.IDPInfo),
+		RefreshToken: args.RefreshToken,
+	}
 }
 
-// NewConfigWithAuthenticator will translate data from client options into a topology config for building non-default deployments.
-// Server and topology options are not honored if a custom deployment is used. It uses a passed in
+// NewAuthenticator returns a [driver.Authenticator] configured with the given
+// credential and HTTP client. It returns nil if cred is nil.
+func NewAuthenticator(cred *options.Credential, httpClient *http.Client) (driver.Authenticator, error) {
+	if cred == nil {
+		return nil, nil
+	}
+
+	var oidcMachineCallback auth.OIDCCallback
+	if cred.OIDCMachineCallback != nil {
+		oidcMachineCallback = func(ctx context.Context, args *driver.OIDCArgs) (*driver.OIDCCredential, error) {
+			cred, err := cred.OIDCMachineCallback(ctx, convertOIDCArgs(args))
+			return (*driver.OIDCCredential)(cred), err
+		}
+	}
+
+	var oidcHumanCallback auth.OIDCCallback
+	if cred.OIDCHumanCallback != nil {
+		oidcHumanCallback = func(ctx context.Context, args *driver.OIDCArgs) (*driver.OIDCCredential, error) {
+			cred, err := cred.OIDCHumanCallback(ctx, convertOIDCArgs(args))
+			return (*driver.OIDCCredential)(cred), err
+		}
+	}
+
+	// Create an authenticator for the client
+	return auth.CreateAuthenticator(
+		cred.AuthMechanism,
+		&auth.Cred{
+			Source:              cred.AuthSource,
+			Username:            cred.Username,
+			Password:            cred.Password,
+			PasswordSet:         cred.PasswordSet,
+			Props:               cred.AuthMechanismProperties,
+			OIDCMachineCallback: oidcMachineCallback,
+			OIDCHumanCallback:   oidcHumanCallback,
+		},
+		httpClient)
+}
+
+// NewConfig will translate data from client options into a topology config for
+// building non-default deployments.
+func NewConfig(co *options.ClientOptions, clock *session.ClusterClock) (*Config, error) {
+	authenticator, err := NewAuthenticator(co.Auth, co.HTTPClient)
+	if err != nil {
+		return nil, fmt.Errorf("error creating authenticator: %w", err)
+	}
+	return NewConfigWithAuthenticator(co, clock, authenticator)
+}
+
+// NewConfigWithAuthenticator will translate data from client options into a
+// topology config for building non-default deployments. Server and topology
+// options are not honored if a custom deployment is used. It uses a passed in
 // authenticator to authenticate the connection.
-func NewConfigWithAuthenticator(co *options.ClientOptions, clock *session.ClusterClock, authenticator driver.Authenticator) (*Config, error) {
+func NewConfigWithAuthenticator(
+	co *options.ClientOptions,
+	clock *session.ClusterClock,
+	authenticator driver.Authenticator,
+) (*Config, error) {
 	var serverAPI *driver.ServerAPIOptions
 
 	if err := co.Validate(); err != nil {
@@ -178,30 +223,8 @@ func NewConfigWithAuthenticator(co *options.ClientOptions, clock *session.Cluste
 	}
 
 	// Handshaker
-	var handshaker = func(driver.Handshaker) driver.Handshaker {
-		return operation.NewHello().AppName(appName).Compressors(comps).ClusterClock(clock).
-			ServerAPI(serverAPI).LoadBalanced(loadBalanced)
-	}
-	// Auth & Database & Password & Username
-	if co.Auth != nil {
-		cred := &auth.Cred{
-			Username:    co.Auth.Username,
-			Password:    co.Auth.Password,
-			PasswordSet: co.Auth.PasswordSet,
-			Props:       co.Auth.AuthMechanismProperties,
-			Source:      co.Auth.AuthSource,
-		}
-		mechanism := co.Auth.AuthMechanism
-
-		if len(cred.Source) == 0 {
-			switch strings.ToUpper(mechanism) {
-			case auth.MongoDBX509, auth.GSSAPI, auth.PLAIN:
-				cred.Source = "$external"
-			default:
-				cred.Source = "admin"
-			}
-		}
-
+	var handshaker func(driver.Handshaker) driver.Handshaker
+	if authenticator != nil {
 		handshakeOpts := &auth.HandshakeOptions{
 			AppName:       appName,
 			Authenticator: authenticator,
@@ -211,9 +234,9 @@ func NewConfigWithAuthenticator(co *options.ClientOptions, clock *session.Cluste
 			ClusterClock:  clock,
 		}
 
-		if mechanism == "" {
+		if co.Auth.AuthMechanism == "" {
 			// Required for SASL mechanism negotiation during handshake
-			handshakeOpts.DBUser = cred.Source + "." + cred.Username
+			handshakeOpts.DBUser = co.Auth.AuthSource + "." + co.Auth.Username
 		}
 		if co.AuthenticateToAnything != nil && *co.AuthenticateToAnything {
 			// Authenticate arbiters
@@ -225,7 +248,17 @@ func NewConfigWithAuthenticator(co *options.ClientOptions, clock *session.Cluste
 		handshaker = func(driver.Handshaker) driver.Handshaker {
 			return auth.Handshaker(nil, handshakeOpts)
 		}
+	} else {
+		handshaker = func(driver.Handshaker) driver.Handshaker {
+			return operation.NewHello().
+				AppName(appName).
+				Compressors(comps).
+				ClusterClock(clock).
+				ServerAPI(serverAPI).
+				LoadBalanced(loadBalanced)
+		}
 	}
+
 	connOpts = append(connOpts, WithHandshaker(handshaker))
 	// ConnectTimeout
 	if co.ConnectTimeout != nil {
